添加一系列接口

+++ src/main/java/com/qianyou/www/pojo/AddHistoryRequest.java
@@ -17,7 +17,7 @@ public class AddHistoryRequest {
     /**
      * 科目id
      */
-    private Long menu;
+    private Integer menu;
 
     /**
      * 分数
